# Conflicts:
#	Depo/Depo.xcodeproj/project.pbxproj
#	Depo/Depo/App/Constants/Constants.swift
#	Depo/Depo/Presentation/UserStories/Settings/SyncContacts/Interactor/Services/ContactsSyncService.swift
#	Depo/Depo/Presentation/UserStories/Settings/SyncContacts/Presenter/SyncContactsPresenter.swift

+++ Depo/Depo/Presentation/UserStories/Grid/FreeAppSpace/Interactor/FreeAppSpaceInteractor.swift
@@ -8,16 +8,22 @@
 
 class FreeAppSpaceInteractor: BaseFilesGreedInteractor {
     
+    var isDeleteRequestRunning = false
+    
     func onDeleteSelectedItems(selectedItems: [WrapData]) {
+        if (isDeleteRequestRunning){
+            return
+        }
+        
+        isDeleteRequestRunning = true
         let uuids = FreeAppSpace.default.getServerUIDSForLocalitem(localItemsArray: selectedItems)
         
         FileService().details(uuids: uuids, success: { [weak self] (objects) in
             //let localFilesForDelete = FreeAppSpace.default.getLocalFiesComaredWithServerObjects(serverObjects: objects, localObjects: selectedItems)
-            
             let array = FreeAppSpace.default.getLocalFiesComaredWithServerObjects(serverObjects: objects, localObjects: selectedItems)
             
             let fileService = WrapItemFileService()
-            fileService.delete(deleteFiles: array, success: {
+            fileService.deleteLocalFiles(deleteFiles: array, success: {
                 
                 FreeAppSpace.default.deleteDeletedLocalPhotos(deletedPhotos: array)
                 
@@ -28,16 +34,20 @@ class FreeAppSpaceInteractor: BaseFilesGreedInteractor {
                 if let service = self_.remoteItems as? FreeAppService {
                     service.clear()
                 }
+                self_.isDeleteRequestRunning = false
                 if let presenter = self_.output as? FreeAppSpacePresenter {
                     DispatchQueue.main.async {
                         presenter.onItemDeleted()
-                        CardsManager.default.stopOperationWithType(type: .freeAppSpace)
-                        CardsManager.default.stopOperationWithType(type: .freeAppSpaceWarning)
+                        if FreeAppSpace.default.getDuplicatesObjects().count == 0{
+                            CardsManager.default.stopOperationWithType(type: .freeAppSpace)
+                            CardsManager.default.stopOperationWithType(type: .freeAppSpaceWarning)
+                        }
                         presenter.goBack()
                     }
                 }
                 
             }, fail: { [weak self] error in
+                self?.isDeleteRequestRunning = false
                 if let presenter = self?.output as? FreeAppSpacePresenter {
                     DispatchQueue.main.async {
                         presenter.reloadData()
@@ -50,6 +60,9 @@ class FreeAppSpaceInteractor: BaseFilesGreedInteractor {
     }
     
     override func reloadItems(_ searchText: String!, sortBy: SortType, sortOrder: SortOrder, newFieldValue: FieldValue?) {
+        if let remoteItems = remoteItems as? FreeAppService {
+            remoteItems.clear()
+        }
         super.reloadItems(searchText, sortBy: sortBy, sortOrder: sortOrder, newFieldValue: newFieldValue)
     }
